Create base Widget class and CommWidget
