+++ pyproject.toml
@@ -15,7 +15,6 @@ authors = [
 keywords = [
     "widgets",
     "jupyter",
-    "ypy",
     "yjs",
 ]
 dependencies = [
