+++ ypywidgets/__init__.py
@@ -1,5 +1,5 @@
-from .ypywidgets import Widget  # noqa
-from .reactive import reactive  # noqa
+from .widget import Widget as Widget
+from .reactive import reactive as reactive
 
 
 __version__ = "0.5.0"
